Print invalid relations
Add the names of the invalid relations when throwing the exception. This
makes it easier to debug any mistakes in the client code or any need to
add more allowed relations.

This should not pose any security risks, as an attacker can already
find out if a relation is invalid by trial and error.